Add support for MUR strategies for Crystallize Time.
Strategy for reference: https://materiaraiding.com/ultimate/fru#crystallize-time

There's two parts to this. First is the MUR priority for red ice/aeros which is just good old-fashioned panto prio. Second part is adding an option to have the aero knocking back the party stand still until after the knockback goes off, a.k.a 'Aero's plant'. Note that this doesn't implement the lineup as described in the raid plan, my understanding is that it's not often done, most people just read the party list, so I've left it out of this change.

To support aeros plant there's a little refactoring to make the aero target location relative to aero source rather than being a fixed spot. This also adds UI elements to select priority (NA or MUR) and whether aero's plant.